Decouple mock data
