+++ src/Emulator/Emulator.tsx
@@ -1,10 +1,10 @@
-import React, { FC } from 'react';
+import React from 'react';
 import styled from '@emotion/styled';
 import { Flex } from '@chakra-ui/core';
 
 import Tabs from './Tabs/Tabs';
 import Visuals from './Visuals';
-import Instructions, { PanelData } from './Instructions';
+import Instructions from './Instructions';
 
 const Container = styled.div`
   margin-left: 1.5vw;
@@ -17,14 +17,23 @@ const Container = styled.div`
   overflow: hidden;
 `;
 
-interface Props {
-  panelData: PanelData;
-}
+const data = {
+  'id': 1,
+  'submoduleNumber': 1,
+  'submoduleName': 'Modifying Topology',
+  'objective': 'Add and remove devices and their connections.',
+  'tasks': [
+    'Add a Host (h3)',
+    'Attach h3 to s2',
+    'Detach h3 from s2',
+    'Remove h2',
+  ],
+};
 
-const Emulator: FC<Props> = ({ panelData }) => {
+const Emulator = () => {
   return (
     <Flex margin='auto' maxWidth='1920px'>
-      <Instructions panelData={panelData} />
+      <Instructions panelData={data} />
       <Container>
         <div>
           <Tabs />
